quick-data/app/requirements.txt
requests
starlette
pydantic
fastapi
uvicorn
kafka-python
urllib3
certifi
elasticsearch
